move short term / legacy snapshot code into child module
